Re-organized armies package out of internal
