+++ armies/armies_test.go
@@ -1,0 +1,14 @@
+package armies_test
+
+import (
+	"github.com/brittonhayes/warhammer-aos/armies"
+	"github.com/brittonhayes/warhammer-aos/internal/handlers"
+)
+
+// Create a new handler
+// and add a function to the route
+func ExampleArmies() {
+	_ = handlers.New().
+		AddRoute("*", armies.Handler()).
+		Build()
+}
