+++ go.mod
@@ -7,4 +7,5 @@ require (
 	github.com/awslabs/aws-lambda-go-api-proxy v0.9.0
 	github.com/gofiber/fiber/v2 v2.1.0
 	github.com/pkg/errors v0.9.1
+	github.com/thedevsaddam/gojsonq v2.3.0+incompatible // indirect
 )
