+++ internal/handlers/response.go
@@ -1,0 +1,8 @@
+package handlers
+
+// Response is the default API response
+// format
+type Response struct {
+	Count int         `json:"count"`
+	Data  interface{} `json:"data"`
+}
